requirement.txt
pandas
mysql.connector
folium
webbrowser
numpy
pandas
matplotlib
seaborn
datetime